requirements.txt
pandas
scikit-learn
joblib
python-dotenv
discord.py
